Pre-Alpha v0.0
Added to .sty:
- problem environment
- solution environment
- solution environment

Added demonstrative test.tex and test.pdf files.

All the ways of using a sty file in TeXlive is pretty janky, except for uploading to CTAN, but it is hard to do that for testing. For development, I just put all the .tex and demo files in the same folder.